Segregated args for buildexe to args_setup.h/.cpp
