+++ buildexe/CMakeLists.txt
@@ -1,5 +1,8 @@
 add_executable(buildexe
     buildexe.cpp
+
+    src/args_setup.cpp
+    include/buildexe/args_setup.h
 )
 target_sources(buildexe PRIVATE
     ../bootstrap/src/build_buildcc.cpp
@@ -11,6 +14,7 @@ target_sources(buildexe PRIVATE
     ../bootstrap/src/build_tpl.cpp
 )
 target_include_directories(buildexe PRIVATE 
+    include
     ../bootstrap/include
 )
 target_link_libraries(buildexe PRIVATE buildcc)
